Simplified the navbar
The navbar now only requires 1 div, rather than 3 nested ones. The
downside to this (AFAICT) is that old versions of IE will no longer get a
fancy gradient. I believe this is a good compromise.

(The container div was used, AFAICT, set set fixed widths in responsive
layouts. Since bootvelcro leaves this up to the user, it too is no
longer required)