Added bootstrap-button.js style checkboxes
The bootstrap-button.js style checkboxes and radio buttons are nice but:
  - require JS
  - aren't semantically correct

So they were removed. This change adds them back in, but removes the JS
dependency and make them more semantically correct. That is, to create
the checkboxes, do something like:

	<label class="btn-checkbox">
		<input type="checkbox">
		<span class="btn">Check me</span>
	</label>

And for a radio button, simply substitute .btn-radio instead.

Note, that this also works with btn-group classes.